binance spot balance done

+++ exchange/model.go
@@ -348,15 +348,15 @@ const (
 
 type KlineDetail struct {
 	ID                  string  `json:"id"`
-	OpenTime            int64   `json:"open_time"`
+	OpenTime            float64 `json:"open_time"`
 	Open                float64 `json:"open"`
 	High                float64 `json:"high"`
 	Low                 float64 `json:"low"`
 	Close               float64 `json:"close"`
 	Volume              float64 `json:"volume"`
-	CloseTime           int64   `json:"close_time"`
+	CloseTime           float64 `json:"close_time"`
 	QuoteAssetVolume    float64 `json:"quote_asset_volume"`
-	TradesCount         int64   `json:"trades_count"`
+	TradesCount         float64 `json:"trades_count"`
 	TakerBuyBaseVolume  float64 `json:"taker_buy_base_volume"`
 	TakerBuyQuoteVolume float64 `json:"taker_buy_quote_volume"`
 }
